Verify controller errors as well
